format data and add multiple series
