+++ frontend/fingertips-frontend/app/chart/health-data.ts
@@ -1,1 +1,10 @@
-// Map data from endpoint
+interface HealthCareData {
+  areaCode: string;
+  healthData: {
+    year: number;
+    count: number;
+    value: number;
+    lowerCi: number;
+    upperCi: number;
+  }[];
+}
