+++ frontend/fingertips-frontend/app/chart/page.tsx
@@ -1,25 +1,14 @@
 import { Chart } from '@/components/pages/chart';
-import { getApiConfiguration } from '@/lib/getApiConfiguration';
 import { connection } from 'next/server';
-import { WeatherForecastApi } from '@/generated-sources/api-client';
 import { getGetHealthDataForAnIndicator200Response } from '@/mock/server/handlers';
 
-interface HealthCareData {
-  areaCode: string;
-  healthData: {
-    year: number;
-    count: number;
-    value: number;
-    lowerCi: number;
-    upperCi: number;
-  }[];
-}
-
-export default async function ChartPage(props: Readonly<{
-  searchParams?: Promise<{
-    indicator?: string;
+export default async function ChartPage(
+  props: Readonly<{
+    searchParams?: Promise<{
+      indicator?: string;
+    }>;
   }>
-}> ) {
+) {
   // We don't want to render this page statically
   await connection();
 
@@ -27,18 +16,10 @@ export default async function ChartPage(props: Readonly<{
   const indicator = searchParams?.indicator ?? '';
 
   const endpoint = getGetHealthDataForAnIndicator200Response();
-  
-  // TODO - Get filtered area code data
-  // const data = endpoint.filter(( item => item.areaCode === areaCode))
- 
+
   const data: HealthCareData[] = endpoint.map((item) => ({
     areaCode: item.areaCode,
-    healthData: item.healthData.map((item ) => ({year: item.year,
-      count: item.count,
-      value: item.value,
-      lowerCi: item.lowerCi,
-      upperCi: item.upperCi
-    }))
+    healthData: item.healthData,
   }));
 
   return <Chart data={data} />;
