+++ frontend/fingertips-frontend/components/organisms/LineChart/index.tsx
@@ -1,17 +1,6 @@
 import Highcharts from 'highcharts';
 import { HighchartsReact } from 'highcharts-react-official';
 
-interface HealthCareData {
-  areaCode: string;
-  healthData: {
-    year: number;
-    count: number;
-    value: number;
-    lowerCi: number;
-    upperCi: number;   
-  }[];
-}
-
 interface LineChartProps {
   data: HealthCareData[];
   title?: string;
@@ -27,33 +16,28 @@ export function LineChart({
   yAxisTitle,
   accessibilityLabel,
 }: Readonly<LineChartProps>) {
-  const value = data.map((item) => (item.healthData.map((point => point.value))));
-  const categories = data.map((item => (item.healthData.map((point => point.year)))));
-
-  console.log('categories', categories);
+  
+  const years = data.map((item) =>
+    item.healthData.map((point) => point.year)
+  );
+  const formatYear = years.flat().map((year) => year.toString())
 
   const lineChartOptions: Highcharts.Options = {
     chart: { type: 'line' },
     title: { text: title },
     xAxis: {
-      categories: categories.map((item => item.toString())),
-      title: { text: xAxisTitle },
+      categories: Array.from(new Set(
+        formatYear)),
+      title: { text: xAxisTitle}
     },
     yAxis: {
       title: { text: yAxisTitle },
     },
-    series: [
-      {
-        type: 'line',
-        name: yAxisTitle,
-        data: value,
-      },
-      {
-        type: 'line',
-        name: yAxisTitle,
-        data: value,
-      },
-    ],
+    series: data.map((item, index) => ({
+      type: 'line',
+      name: `AreaCode ${item.areaCode}`, 
+      data: item.healthData.map((point) => point.value),
+    })),
     accessibility: {
       enabled: false,
       description: accessibilityLabel,
