+++ frontend/fingertips-frontend/components/organisms/LineChartTable/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { Table } from 'govuk-react';
-import { WeatherForecast } from '@/generated-sources/api-client';
 
 interface HealthCareData {
   areaCode: string;
@@ -10,7 +9,7 @@ interface HealthCareData {
     count: number;
     value: number;
     lowerCi: number;
-    upperCi: number;   
+    upperCi: number;
   };
 }
 
@@ -33,9 +32,7 @@ export function LineChartTable({ data, headings }: Readonly<TableProps>) {
       >
         {data.map((item) => (
           <Table.Row key={`${item.healthData?.year}`}>
-            <Table.Cell>
-              {item.healthData?.year}
-            </Table.Cell>
+            <Table.Cell>{item.healthData?.year}</Table.Cell>
             <Table.Cell numeric>{item.healthData?.value}</Table.Cell>
             <Table.Cell numeric>{item.healthData?.count}</Table.Cell>
             <Table.Cell numeric>{item.healthData?.lowerCi}</Table.Cell>
